requirements.txt
streamlit
tweepy
gspread
pandas
google-auth
